WeatherData/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
kafka-python-2.0.2
urllib3==1.26.12
psycopg2-2.9.5
sqlalchemy-1.4.45